Switch version 3.4 to 3.5

+++ LSN-Helper.lua
@@ -2,7 +2,7 @@ script_name('LSN-Helper')
 script_description('Los Santos News Helper (LSNH) for special project MyHome RP')
 script_author('kyrtion#7310')
 script_properties('work-in-pause')
-script_version('3.4')
+script_version('3.5')
 
 require 'lib.moonloader'
 local dlstatus = require('moonloader').download_status
@@ -50,12 +50,14 @@ local new, str, sizeof = imgui.new, ffi.string, ffi.sizeof
 local renderWindow = new.bool(false)
 
 local adInput = new.char[256]('')
-local inputSearch = new.char[256]('')
+local searchInput = new.char[256]('')
+
+--local adNick, adPrice, adText = 'Awe Some[123]', '138$', 'awesome text, dude'
 local adNick, adPrice, adText = '', '', ''
+local notAdNick = false
 local confirm = false
 local block = false
 local copying = false
-local autoFocus = false
 local hex = '0xEEDC82'
 
 local update_state = false
@@ -65,11 +67,19 @@ local lockFailed = false
 local newVersion = 'None'
 local oldVersion = 'None'
 
+--! origin/master
 local update_url = 'https://raw.githubusercontent.com/kyrtion/LSNHelper_mhrp/master/version_lsn.ini'
 local update_path = getWorkingDirectory() .. '/update_lsn.ini'
 local script_vers = tostring(thisScript().version)
 local script_url = 'https://github.com/kyrtion/LSNHelper_mhrp/blob/master/LSN-Helper.lua?raw=true'
 local script_path = thisScript().path
+
+-- --! origin/beta
+-- local update_url = 'https://raw.githubusercontent.com/kyrtion/LSNHelper_mhrp/beta/version_lsn.ini'
+-- local update_path = getWorkingDirectory() .. '/update_lsn.ini'
+-- local script_vers = tostring(thisScript().version)
+-- local script_url = 'https://github.com/kyrtion/LSNHelper_mhrp/blob/beta/LSN-Helper.lua?raw=true'
+-- local script_path = thisScript().path
 
 function send(result) sampAddChatMessage('LSNH » '.. result, 0xEEDC82) end
 
@@ -82,7 +92,7 @@ local newFrame = imgui.OnFrame(
 		local sizeX, sizeY = 700, 340
 		imgui.SetNextWindowPos(imgui.ImVec2(resX / 2, resY / 2), imgui.Cond.FirstUseEver, imgui.ImVec2(0.5, 0.5))
 		imgui.SetNextWindowSize(imgui.ImVec2(sizeX, sizeY * 1.04))
-		imgui.Begin(u8'Публикация oбъявления | LSN-Helper '..thisScript().version, nil, imgui.WindowFlags.NoMove + imgui.WindowFlags.NoResize + imgui.WindowFlags.NoCollapse + imgui.WindowFlags.AlwaysAutoResize)
+		imgui.Begin(u8'Публикация oбъявления | LSN-Helper '..thisScript().version, nil, imgui.WindowFlags.NoMove + imgui.WindowFlags.NoResize + imgui.WindowFlags.NoCollapse)
 
 		imgui.SetCursorPos(imgui.ImVec2(20, 40)); imgui.TextColoredRGB('Отправитель:'); imgui.SameLine((sizeX - 15) / 2 + 10); imgui.TextColoredRGB('Цена:')
 		imgui.SetCursorPos(imgui.ImVec2(20, 60));
@@ -116,7 +126,7 @@ local newFrame = imgui.OnFrame(
 
 		imgui.SetCursorPos(imgui.ImVec2(20, sizeY - 110))
 		imgui.PushItemWidth(sizeX - 40);
-
+		
 		if imgui.IsWindowAppearing() then imgui.SetKeyboardFocusHere(-1) end
 		imgui.PushAllowKeyboardFocus(false)
 		imgui.InputText(u8'##adInput', adInput, sizeof(adInput))
@@ -127,8 +137,8 @@ local newFrame = imgui.OnFrame(
 		imgui.PushStyleColor(imgui.Col.ButtonHovered, imgui.ImVec4(0.34, 0.42, 0.51, 0.9))
 		imgui.PushStyleColor(imgui.Col.ButtonActive, imgui.ImVec4(0.34, 0.42, 0.51, 0.8))
 		if imgui.Button(u8'Передать в /rb', imgui.ImVec2((sizeX - 42) / 2 , 25)) then
-			if (u8:decode(str(adInput))) == (nil or '') then
-				send('В тексте пусто, зачем отправлять?', -1)
+			if (u8:decode(adText)) == (nil or '') then
+				send('В тексте пусто, зачем сообщать?', -1)
 			else
 				sampSendChat('/rb '.. adNick .. ' (' .. adPrice .. '): '.. adText)
 			end
@@ -140,60 +150,119 @@ local newFrame = imgui.OnFrame(
 			imgui.OpenPopup(u8'Поиск')
 		end
 
-		if imgui.BeginPopupModal(u8'Поиск', _, imgui.WindowFlags.NoMove + imgui.WindowFlags.NoResize + imgui.WindowFlags.NoCollapse + imgui.WindowFlags.AlwaysAutoResize) then
-			local pSize = imgui.ImVec2(770, 340)
+		if imgui.BeginPopupModal(u8'Поиск', _, imgui.WindowFlags.NoMove + imgui.WindowFlags.NoResize + imgui.WindowFlags.NoCollapse) then
+			local pSize = imgui.ImVec2(770, 400)
 			imgui.SetWindowSizeVec2(pSize)
-	
+			
 			imgui.SetCursorPos(imgui.ImVec2(20, 45))
 			imgui.Text(u8'Текст:') imgui.SameLine(70) imgui.TextColoredRGB(adText)
-				
-			imgui.SetCursorPos(imgui.ImVec2(20, 70)) imgui.Text(u8'Поиск:') imgui.SameLine(70)
 			
-			imgui.SetCursorPos(imgui.ImVec2(70, 67))
-			imgui.PushItemWidth(pSize.x - 110)
+			imgui.SetCursorPos(imgui.ImVec2(20, 70)) imgui.Text(u8'Поиск:')
+			
+			imgui.SetCursorPos(imgui.ImVec2(70, 66))
+			imgui.PushItemWidth(pSize.x/2 + 294)
 
 			if imgui.IsWindowAppearing() then imgui.SetKeyboardFocusHere(-1) end
 			imgui.PushAllowKeyboardFocus(false)
-			if imgui.InputText('##inputSearch', inputSearch, sizeof(inputSearch)) then str(inputSearch):find(str(inputSearch):gsub("%p", "%%%1")) end
+			if imgui.InputText('##searchInput', searchInput, sizeof(searchInput)) then str(searchInput):find(str(searchInput):gsub("%p", "%%%1")) end
 			imgui.PopAllowKeyboardFocus()
 			imgui.PopItemWidth()
 			
-			imgui.SetCursorPos(imgui.ImVec2(20, 108))
+			imgui.SetCursorPos(imgui.ImVec2(21, 108))
 			imgui.Text(u8'Результаты:')
-	
-			imgui.SetCursorPos(imgui.ImVec2(19, 130))
-			imgui.BeginChild('ChildWindowsS', imgui.ImVec2(pSize.x/2 + 325, pSize.y/2 - 6), true)
 			
+			imgui.SetCursorPos(imgui.ImVec2(20, 130))
+			imgui.BeginChild('ChildWindowsS', imgui.ImVec2(pSize.x/2 + 344, pSize.y/2 + 12), true)
+			--imgui.Separator()
+			local size = imgui.GetWindowSize()
+            local ColumnsSize = {
+                [1] = 59,
+                [2] = 90,
+                [3] = 90,
+                [4] = size.x - 75 - 90 - 40 - 30,
+                [5] = 24
+            }
+
 			for i=1, #adList do
-				if string.len(str(inputSearch)) ~= 0 then
-					if string.find(u8(adList[i]), str(inputSearch), 1, true) then
-						if imgui.Button(u8' > ##'..tostring(i)) then
+				if string.len(str(searchInput)) ~= 0 then
+					if string.find(u8(adList[i]), str(searchInput), 1, true) then
+						imgui.Columns(2)
+						if imgui.Selectable(u8'  >  ##'..tonumber(i), true, _, imgui.ImVec2(20, 15)) then
 							imgui.StrCopy(adInput, u8(adList[i]))
 							imgui.CloseCurrentPopup()
-							inputSearch = new.char[256]('')
+							searchInput = new.char[256]('')
 						end
-						imgui.SameLine(36)
+						imgui.SameLine(32)
+						if imgui.Selectable(u8' RB ##'..tonumber(i), true, _, imgui.ImVec2(22, 15)) then
+							sampSendChat('/rb >> '..adList[i])
+						end
+						imgui.SetColumnWidth(0, ColumnsSize[1])
+						imgui.NextColumn()
 						imgui.Text(u8(adList[i]))
+						imgui.NextColumn()
 						imgui.Separator()
 					end
 				else
-					if imgui.Button(u8' > ##'..tostring(i)) then
+					imgui.Columns(2)
+					if imgui.Selectable(u8'  >  ##'..tonumber(i), true, _, imgui.ImVec2(20, 15)) then
 						imgui.StrCopy(adInput, u8(adList[i]))
-						inputSearch = new.char[256]('')
 						imgui.CloseCurrentPopup()
+						searchInput = new.char[256]('')
 					end
-					imgui.SameLine(36)
+					imgui.SameLine(32)
+					if imgui.Selectable(u8' RB ##'..tonumber(i), true, _, imgui.ImVec2(22, 15)) then
+						sampSendChat('/rb >> '..adList[i])
+					end
+					imgui.SetColumnWidth(0, ColumnsSize[1])
+					imgui.NextColumn()
 					imgui.Text(u8(adList[i]))
+					imgui.NextColumn()
 					imgui.Separator()
 				end
 			end
+			-- будет второй вариант для результаты поисков
+			--[[for i=1, #adList do
+				if string.len(str(searchInput)) ~= 0 then
+					if string.find(u8(adList[i]), str(searchInput), 1, true) then
+						
+						if imgui.Button('>##'..tostring(i), imgui.ImVec2(22, 24)) then
+							imgui.StrCopy(adInput, u8(adList[i]))
+							imgui.CloseCurrentPopup()
+							searchInput = new.char[256]('')
+						end
+						imgui.SameLine()
+						if imgui.Button('RB##'..tostring(i)) then
+							sampSendChat('/rb >> '..adList[i])
+						end
+						imgui.SameLine()
+						imgui.Text(u8(adList[i]))
+						if i ~= #adList then imgui.Separator() end
+
+					end
+				else
+
+					if imgui.Button('>##'..tostring(i), imgui.ImVec2(22, 24)) then
+						imgui.StrCopy(adInput, u8(adList[i]))
+						imgui.CloseCurrentPopup()
+						searchInput = new.char[256]('')
+					end
+					imgui.SameLine()
+					if imgui.Button('RB##'..tostring(i)) then
+						sampSendChat('/rb >> '..adList[i])
+					end
+					imgui.SameLine()
+					imgui.Text(u8(adList[i]))
+					if i ~= #adList then imgui.Separator() end
+
+				end
+			end]]
 
 			imgui.EndChild()
 
 
-			imgui.SetCursorPos(imgui.ImVec2(5, pSize.y - 20))
-			if imgui.Button(u8'Закрыть', imgui.ImVec2(pSize.x - 30, 25)) then
-				inputSearch = new.char[256]('')
+			imgui.SetCursorPos(imgui.ImVec2(20, pSize.y - 42))
+			if imgui.Button(u8'Закрыть', imgui.ImVec2(pSize.x/2 + 344, 25)) then
+				searchInput = new.char[256]('')
 				imgui.CloseCurrentPopup()
 			end
 	
@@ -212,7 +281,7 @@ local newFrame = imgui.OnFrame(
 				send('В тексте пусто, зачем отправлять?', -1)
 
 			elseif Char and Char ~= ('$' or ',' or '/' or '>' or '<' or '-' or '=' or '+' or '_' or "'" or '"') then
-				sampSendDialogResponse(1536,1,0,(u8:decode(str(adInput))))
+				sampSendDialogResponse(1536, 1, 0, (u8:decode(str(adInput))))
 				renderWindow[0] = false
 				confirm = true
 
@@ -224,7 +293,7 @@ local newFrame = imgui.OnFrame(
 				adNick, adPrice, adText = '', '', ''
 			
 			else
-				send('В конце знаки препинание не ставил!')
+				send('Вы не ставили в конце знаки препинание')
 			end
 		end
 		imgui.PopStyleColor(3)
@@ -235,9 +304,9 @@ local newFrame = imgui.OnFrame(
 		imgui.SameLine((sizeX - 17) / 2 + 10)
 		if imgui.Button(u8'Отклонить', imgui.ImVec2((sizeX - 42) / 2 , 25)) then
 			if (u8:decode(str(adInput))) == (nil or '') then
-				send('Вы ничего не ввели в поле ввода', -1)
+				send('Вы не указали причину в поле', -1)
 			else
-				sampSendDialogResponse(1536,0,0,(u8:decode(str(adInput))))
+				sampSendDialogResponse(1536, 0, 0, (u8:decode(str(adInput))))
 				
 				renderWindow[0] = false
 				copying = false
@@ -269,16 +338,15 @@ function main()
 	send('Скрипт успешно загружено. Версия: '..thisScript().version)
 	print(); print('Script LSN-Helper '..thisScript().version..' loaded - Discord: kyrtion#7310')
 
-	--! debug window (dont use)
-	-- sampRegisterChatCommand('dia', function()
+	-- --! debug window (dont use)
+	-- sampRegisterChatCommand('ef', function()
 	-- 	renderWindow[0] = not renderWindow[0]
 	-- 	imgui.StrCopy(adInput, u8(adText))
-	-- 	autoFocus = true
 	-- end)
 
 	sampRegisterChatCommand('verify', function()
 		if lockVerify then
-			if renderWindow[0] and sampIsDialogActive() then				
+			if renderWindow[0] or sampIsDialogActive() then				
 				send('Закройте диалог и снова вводите /verify')
 			else
 				checkVerify = true
@@ -291,15 +359,18 @@ function main()
 	downloadUrlToFile(update_url, update_path, function(id, status)
 		if status == dlstatus.STATUS_ENDDOWNLOADDATA then
 			updateIni = inicfg.load(nil, update_path)
-			newVersion = tostring(updateIni.info.version):gsub('"', '')
-			oldVersion = tostring(thisScript().version)
-			--sampAddChatMessage(newVersion..' -> '..oldVersion, -1)
-			if newVersion ~= oldVersion then
-				send('Есть обновление! Версия: '..newVersion..'. Чтобы обновить вводите /verify', -1)
-				update_state = true
-				lockVerify = true
+			if updateIni ~= nil or updateIni.info.version ~= nil then
+				newVersion = tostring(updateIni.info.version):gsub('"', '')
+				oldVersion = tostring(thisScript().version)
+				if newVersion ~= oldVersion then
+					send('Есть обновление! Версия: '..newVersion..'. Чтобы обновить вводите /verify', -1)
+					update_state = true
+					lockVerify = true
+				end
+				os.remove(update_path)
+			else
+				send('Невозможно проверить наличие обновление. Нажмите CTRL + R чтобы перезагрузить')
 			end
-			os.remove(update_path)
 		end
 	end)
 	
@@ -318,8 +389,6 @@ function main()
 	end
 end
 
-
-
 function onWindowMessage(msg, wparam, lparam)
 	if msg == 0x100 or msg == 0x101 then
 		if (wparam == VK_ESCAPE and (renderWindow[0])) and not isPauseMenuActive() and not sampIsChatInputActive() and not isSampfuncsConsoleActive() then
@@ -330,23 +399,26 @@ end
 
 function sampev.onShowDialog(id, style, title, button1, button2, text)
 	if id == 1536 and title == '{6333FF}Публикация объявления' then
-		--local adN = ''
-		--adN = ( text:match('%{ffffff%}Отправитель%: %{7FFF00%}(%w+ %w+)') ):gsub("\n", "")
+		if notAdNick then
+			adNick = ( text:match('%{ffffff%}Отправитель%: %{7FFF00%}(%w+ %w+)') ):gsub("\n", "")
+			notAdNick = false
+			send('Вы не обновили актуальный диалог для редактирования объявление, ID отправителя не будет показан.')
+			send('В следующий раз когда появится сообщение /edit, вводите команды только без диалога!')
+		end
 		adText = ( text:match('%{ffffff%}Текст%:%{7FFF00%} (.*)%{ffffff%}') ):gsub("\n", "")
 		adPrice = ( text:match('%{ffffff%}Цена%:%{7FFF00%} (.*)%{FFFFFF%}') ):gsub("\n", "")
 		renderWindow[0] = true
-		autoFocus = true
 		if editList[adText] == nil then
 			imgui.StrCopy(adInput, u8(adText))
 			copying = false
-		else	
+		else
 			imgui.StrCopy(adInput, u8(editList[adText]))
 			copying = true
 		end
 		return false
 
 	elseif id == 1537 and title == '{6333FF}Публикация объявления: {ffffff}Подтверждение' and confirm then
-		sampSendDialogResponse(1537,1,0,0)
+		sampSendDialogResponse(1537, 1, 0, 0)
 		confirm = false
 		return false
 	end
@@ -354,7 +426,12 @@ end
 
 function sampev.onSendDialogResponse(dialogId, button, listboxId, input)
 	if dialogId == 1000 and button == 1 then
-		local fr = ''; fr, adNick = input:match('(%d+)%. (.*)')
+		if input:find('-') then
+			notAdNick = true
+		else
+			local fr = ''; fr, adNick = input:match('(%d+)%. (.*)')
+			notAdNick = false
+		end
 	end
 end
 
@@ -516,7 +593,7 @@ function imgui.DarkTheme()
 	--==[ ALIGN ]==--
 	imgui.GetStyle().WindowTitleAlign = imgui.ImVec2(0.5, 0.5)
 	imgui.GetStyle().ButtonTextAlign = imgui.ImVec2(0.5, 0.5)
-	imgui.GetStyle().SelectableTextAlign = imgui.ImVec2(0.5, 0.5)
+	imgui.GetStyle().SelectableTextAlign = imgui.ImVec2(0, 0)
 	
 	--==[ COLORS ]==--
 	imgui.GetStyle().Colors[imgui.Col.Text]                   = imgui.ImVec4(1.00, 1.00, 1.00, 1.00)
